aplicacao do exception para update e ajuste no nome do atributo nome na classe user

+++ src/main/java/com/example/projetospring/services_funcoes/UserService.java
@@ -1,9 +1,12 @@
 package com.example.projetospring.services_funcoes;
 
 import com.example.projetospring.entities.User;
+import com.example.projetospring.services_funcoes.exceptions.DatabaseException;
 import com.example.projetospring.services_funcoes.exceptions.ResourceNotFoundException;
 import com.example.projetospring.repositories_acesso_banco.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -28,19 +31,38 @@ public class UserService {
     }
 
     public void deleteById(Long id) {
-        userRepository.deleteById(id);
 
+        if (!userRepository.existsById(id)) {
+            throw new ResourceNotFoundException(id);
+        }
+
+        try {
+            userRepository.deleteById(id);
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Não é possível deletar: violação de integridade.");
+        }
     }
 
     public User update(Long id, User user) {
-        User entity = userRepository.getReferenceById(id);
-       updateData(entity, user);
 
+        User entity = userRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException(id));
+
+        updateData(entity, user);
         return userRepository.save(entity);
     }
 
+
+    /*public User update(Long id, User user) {
+
+            User entity = userRepository.getReferenceById(id);
+            updateData(entity, user);
+            return userRepository.save(entity);
+
+    }*/
+
     private void updateData(User entity, User user) {
-        entity.setNome(user.getNome());
+        entity.setname(user.getname());
         entity.setEmail(user.getEmail());
         entity.setPhone(user.getPhone());
